refactor: 컴포넌트/스타일 파일 이름 변경
