+++ src/component/CardDetailPage/CardDetailForm/CardDetailForm.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import CardDetailForm from "./index";
+import CardDetailForm from "./CardDetailForm";
 
 const meta = {
   component: CardDetailForm,
