+++ src/component/CardDetailPage/CardDetailForm/CardOwnerNameInput/CardOwnerNameInput.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import St from "./styled";
+import St from "./CardOwnerNameInputStyled";
 
 type CardOwnerNameInputProps = {
   changeCardOwnerName: (e: React.ChangeEvent<HTMLInputElement>) => void;
