+++ src/component/CardDetailPage/CardDetailHeader/CardDetailHeaderStyled.tsx
