+++ src/component/CardDetailPage/CardDetailPageStyled.tsx
