+++ src/component/CardDetailView/CardDetailViewStyled.tsx
