+++ src/component/CardListPage/CardList/CardAddButton/CardAddButton.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import St from "./styled";
+import St from "./CardAddButtonStyled";
 
 function CardAddButton() {
   const navigate = useNavigate();
