+++ src/component/CardListPage/CardListHeader/CardListHeader.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import St from "./styled";
+import St from "./CardListHeaderStyled";
 
 function CardListHeader() {
   return <St.Title>보유 카드</St.Title>;
