+++ src/component/CardListPage/CardListPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import St from "./styled";
-import CardListHeader from "./CardListHeader";
-import CardList from "./CardList";
+import St from "./CardListPageStyled";
+import CardListHeader from "./CardListHeader/CardListHeader";
+import CardList from "./CardList/CardList";
 import { CreditCard } from "../../types/card";
 
 type CardListPageProps = {
